requirements.txt
Django==5.2.7
google-generativeai==0.8.5
python-dotenv==1.2.1
gunicorn==22.0.0 # For production deployment on GCP